backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
opencv-python==4.9.0.80
httpx==0.27.0
